resume_scorer/requirements.txt
python-docx
PyMuPDF
openai
tabulate